+++ packages/config/src/utils/extensions/auto-install-extensions.ts
@@ -18,7 +18,6 @@ interface AutoInstallOptions {
   accountId: string
   token: string
   cwd: string
-  accounts: any
   integrations: IntegrationResponse[]
   offline: boolean
   testOpts: any
@@ -32,20 +31,66 @@ export async function handleAutoInstallExtensions({
   accountId,
   token,
   cwd,
-  accounts,
   integrations,
   offline,
   testOpts = {},
   mode,
   extensionApiBaseUrl,
 }: AutoInstallOptions) {
-  if (!featureFlags?.auto_install_required_extensions || !accountId || !siteId || !token || !cwd || offline) {
+  if (!featureFlags?.auto_install_required_extensions) {
     return integrations
   }
-  const account = accounts?.find((account: any) => account.id === accountId)
-  if (!account) {
+  if (!accountId) {
+    console.error("Failed to auto install extension(s): Missing 'accountId'", {
+      accountId,
+      siteId,
+      cwd,
+      offline,
+      mode,
+    })
     return integrations
   }
+  if (!siteId) {
+    console.error("Failed to auto install extension(s): Missing 'siteId'", {
+      accountId,
+      siteId,
+      cwd,
+      offline,
+      mode,
+    })
+    return integrations
+  }
+  if (!token) {
+    console.error("Failed to auto install extension(s): Missing 'token'", {
+      accountId,
+      siteId,
+      cwd,
+      offline,
+      mode,
+    })
+    return integrations
+  }
+  if (!cwd) {
+    console.error("Failed to auto install extension(s): Missing 'cwd'", {
+      accountId,
+      siteId,
+      cwd,
+      offline,
+      mode,
+    })
+    return integrations
+  }
+  if (offline) {
+    console.error("Failed to auto install extension(s): Running as 'offline'", {
+      accountId,
+      siteId,
+      cwd,
+      offline,
+      mode,
+    })
+    return integrations
+  }
+
   try {
     const packageJson = getPackageJSON(cwd)
     if (
@@ -67,7 +112,7 @@ export async function handleAutoInstallExtensions({
     const results = await Promise.all(
       extensionsToInstall.map(async (ext) => {
         console.log(
-          `Installing extension "${ext.slug}" on team "${account.name}" required by package(s): "${ext.packages.join(
+          `Installing extension "${ext.slug}" on team "${accountId}" required by package(s): "${ext.packages.join(
             '",',
           )}"`,
         )